+++ src/lib.rs
@@ -1538,6 +1538,7 @@ mod tests {
     include!("tests/pg_test_decomposed_chain.rs");
     include!("tests/pg_test_ps3.rs");
     include!("tests/pg_test_ps9.rs");
+    include!("tests/pg_test_ps10.rs");
 }
 
 /// This module is required by `cargo pgrx test` invocations.